Adding `multiple-value-or` macro
Evaluates FORM arguments one at time, until the first value returned
by one of the forms is true.  It then returns all the values returned
by evaluating that form.  If none of the forms return a true first
value, we return the values returned by the last form.

This is required in future changes to SEL.