Update temporary file macros to match uiop/stream
Updating the temporary file and directory macros
to match the name and signature of those in
uiop/stream.  This is the first phase in replacing
these home-rolled macros with those in uiop/stream.
In the second phase, we wil remove those macros
duplicated in uiop/stream from gt/filesystem
and export the uiop/stream versions instead.